junit offline mode working but still have 2 bugs to fix
