+++ com.gzoltar.core/src/main/java/com/gzoltar/core/listeners/JUnitListener.java
@@ -16,53 +16,89 @@
  */
 package com.gzoltar.core.listeners;
 
+import com.gzoltar.core.model.TransactionOutcome;
+import com.gzoltar.core.runtime.Collector;
+import org.junit.platform.engine.TestExecutionResult;
+import org.junit.platform.launcher.TestIdentifier;
+import org.junit.platform.launcher.TestPlan;
 import org.junit.runner.Description;
 import org.junit.runner.Result;
 import org.junit.runner.notification.Failure;
+
+import java.io.PrintWriter;
+import java.io.StringWriter;
+import java.util.Optional;
 
 /**
  * JUnit listener.
  */
 public final class JUnitListener extends Listener {
 
+  public static final String TEST_CLASS_NAME_SEPARATOR = "#";
+
+  private boolean hasFailed = false;
+
+  private long startTime;
+
+  protected String stackTrace;
+
+
+
+  /*
+   * Called when the execution of the TestPlan has finished,
+   * after all tests have been executed.
+   *
+   * Note: This method corresponds to the JUnit 4 method onRunFinish
+   */
   @Override
-  public void testRunStarted(final Description description) {
-    super.onRunStart();
+  public void testPlanExecutionFinished(TestPlan testPlan) {
+    super.testPlanExecutionFinished(testPlan);
+    Collector.instance().endSession();
   }
 
+
+  /*
+   * Called when the execution of a leaf or subtree
+   * of the TestPlan is about to be started.
+   *
+   * Note: This method corresponds to the JUnit 4 method onTestStart
+   */
   @Override
-  public void testRunFinished(final Result result) {
-    super.onRunFinish();
+  public void executionStarted(TestIdentifier testIdentifier) {
+    super.executionStarted(testIdentifier);
+    this.hasFailed = false;
+    this.startTime = System.nanoTime();
+    this.stackTrace = "";
+
   }
 
+  /**
+   * Called when the execution of a leaf or subtree
+   * of the TestPlan has finished, regardless of the outcome.
+   * @param testIdentifier: the TestIdentifier of the test or container that was executed
+   * @param testExecutionResult: the result of the test execution
+   *
+   * Note: This method corresponds to the JUnit 4 method onTestFinished and onTestFailure
+   */
   @Override
-  public void testStarted(final Description description) {
-    super.onTestStart();
-  }
+  public void executionFinished(TestIdentifier testIdentifier, TestExecutionResult testExecutionResult) {
+    super.executionFinished(testIdentifier,testExecutionResult);
 
-  @Override
-  public void testFinished(final Description description) {
-    super.onTestFinish(this.getName(description));
-  }
+    if (testIdentifier.isTest()){
+      if (testExecutionResult.getStatus() == TestExecutionResult.Status.FAILED) {
+        this.hasFailed = true;
+        //this.stackTrace = trace;
+        Optional<Throwable> throwableOp = testExecutionResult.getThrowable();
+        if (throwableOp.isPresent()) {
+          Throwable exception= throwableOp.get();
+          this.stackTrace = traceToString(exception);
+        }
+      }
+    }
 
-  @Override
-  public void testFailure(final Failure failure) {
-    super.onTestFailure(failure.getTrace());
-  }
-
-  @Override
-  public void testAssumptionFailure(final Failure failure) {
-    // an assumption failure is not propagated to org.junit.runner.Result
-  }
-
-  @Override
-  public void testIgnored(final Description description) {
-    super.onTestSkipped();
-  }
-
-  private String getName(final Description description) {
-    return description.getClassName() + Listener.TEST_CLASS_NAME_SEPARATOR
-        + description.getMethodName();
+    Collector.instance().endTransaction(testIdentifier.getDisplayName(),
+            this.hasFailed ? TransactionOutcome.FAIL : TransactionOutcome.PASS,
+            System.nanoTime() - this.startTime, this.stackTrace);
   }
 
 }
