+++ com.gzoltar.core/src/main/java/com/gzoltar/core/listeners/Listener.java
@@ -18,6 +18,8 @@ package com.gzoltar.core.listeners;
 
 import java.io.PrintWriter;
 import java.io.StringWriter;
+
+import org.junit.platform.launcher.listeners.SummaryGeneratingListener;
 import org.junit.runner.notification.RunListener;
 import com.gzoltar.core.model.TransactionOutcome;
 import com.gzoltar.core.runtime.Collector;
@@ -25,7 +27,7 @@ import com.gzoltar.core.runtime.Collector;
 /**
  * Test listener.
  */
-public class Listener extends RunListener {
+public class Listener extends SummaryGeneratingListener {
 
   public static final String TEST_CLASS_NAME_SEPARATOR = "#";
